Bump ruby to a buildpack supported version

+++ app/assets/javascripts/discourse/controllers/user.js.es6
@@ -6,16 +6,28 @@ import User from 'discourse/models/user';
 export default Ember.Controller.extend(CanCheckEmails, {
   indexStream: false,
   userActionType: null,
-  needs: ['user-notifications', 'user-topics-list'],
+  application: Ember.inject.controller(),
+  userNotifications: Ember.inject.controller('user-notifications'),
+  currentPath: Ember.computed.alias('application.currentPath'),
 
   @computed("content.username")
   viewingSelf(username) {
     return username === User.currentProp('username');
   },
 
+  @computed('model.profileBackground')
+  hasProfileBackground(background) {
+    return !Ember.isEmpty(background.toString());
+  },
+
   @computed('indexStream', 'viewingSelf', 'forceExpand')
   collapsedInfo(indexStream, viewingSelf, forceExpand){
     return (!indexStream || viewingSelf) && !forceExpand;
+  },
+
+  @computed('model.isSuspended', 'currentUser.staff')
+  isNotSuspendedOrIsStaff(isSuspended, isStaff) {
+    return !isSuspended || isStaff;
   },
 
   linkWebsite: Em.computed.not('model.isBasic'),
@@ -40,7 +52,12 @@ export default Ember.Controller.extend(CanCheckEmails, {
     return viewingSelf || staff;
   },
 
-  @computed("content.badge_count")
+  @computed('model.name')
+  nameFirst(name) {
+    return !this.get('siteSettings.prioritize_username_in_ux') && name && name.trim().length > 0;
+  },
+
+  @computed("model.badge_count")
   showBadges(badgeCount) {
     return Discourse.SiteSettings.enable_badges && badgeCount > 0;
   },
@@ -69,7 +86,8 @@ export default Ember.Controller.extend(CanCheckEmails, {
     const siteUserFields = this.site.get('user_fields');
     if (!Ember.isEmpty(siteUserFields)) {
       const userFields = this.get('model.user_fields');
-      return siteUserFields.filterProperty('show_on_profile', true).sortBy('position').map(field => {
+      return siteUserFields.filterBy('show_on_profile', true).sortBy('position').map(field => {
+        field.dasherized_name = field.get('name').dasherize();
         const value = userFields ? userFields[field.get('id').toString()] : null;
         return Ember.isEmpty(value) ? null : Ember.Object.create({ value, field });
       }).compact();
@@ -84,8 +102,7 @@ export default Ember.Controller.extend(CanCheckEmails, {
     adminDelete() {
       // I really want this deferred, don't want to bring in all this code till used
       const AdminUser = require('admin/models/admin-user').default;
-      AdminUser.find(this.get('model.username').toLowerCase())
-                         .then(user => user.destroy({deletePosts: true}));
+      AdminUser.find(this.get('model.id')).then(user => user.destroy({deletePosts: true}));
     },
 
   }
